start to split up model file structure
