+++ src/js/reducers/reducer_fallingMagnitude.js
@@ -1,5 +1,6 @@
 import { FALL_TILES, RESET_MAGNITUDE } from '../actions/actionTypes.js'
-import { falseBoard, mapFallingTiles, leavingBoard, removeSeeds } from '../model/model.js'
+import { falseBoard, leavingBoard, removeSeeds } from '../model/model.js'
+import { mapFallingTiles } from '../model/mapFallingTiles.js'
 
 const defaultState = falseBoard()
 export default (state = defaultState, action) => {
