+++ src/js/selectors/selector_fallingMagnitude.js
@@ -1,6 +1,7 @@
 import { createSelector } from 'reselect'
 import { STOP_DRAG, RESET_MAGNITUDE } from '../actions/actionTypes.js'
-import { falseBoard, mapFallingTiles, leavingBoard } from '../model/model.js'
+import { falseBoard, leavingBoard } from '../model/model.js'
+import { mapFallingTiles } from '../model/mapFallingTiles.js'
 
 const board = (state) => state.board
 const moveArray = (state) => state.moves.moveArray
@@ -9,7 +10,7 @@ const isLeavingArray = (state) => state.isLeavingArray
 const defaultState = falseBoard()
 const fallingMagnitudeArray = createSelector(
   [ board, moveArray, isLeavingArray ],
-  (board, moveArray, isLeavingArray) => { // eslint-disable-line 
+  (board, moveArray, isLeavingArray) => { // eslint-disable-line
     switch (isLeavingArray) {
     case STOP_DRAG:
       return mapFallingTiles(leavingBoard(moveArray, board))
